requirements.txt
dj-database-url==0.4.1
Django==1.9.5
django-bootstrap-form==3.2
django-daterange-filter==1.2.0
django-extensions==1.6.1
django-localflavor==1.2
django-widget-tweaks==1.4.1
names==0.3.0
python-decouple==3.0
pytz==2016.4
selenium==2.53.2
six==1.10.0
